refactor load_backbone.py into backbone.py
